I moved calculate_distances to a new folder, packet_transfer. Dependencies were renamed.
